bench: use runtime.nanotime()
Avoids a tiny extra overhead of time.Now(), but the later is also good.
The TSC option should be the lightest, but the repo used uses CPUID as a fence which causes significant overhead in this case (tradeoff for better accuracy). A plain RDTSC(P) might be the best option in the future.